Enter-Object (with) function (#1)
Add the Enter-Object function for accessing the properties of an object or items of a hashtable as variables within a new scope without needing to refer to the containing object. This makes it easy to splat a nested object without temporary variable assignments.

It has an alias of 'with' to provide a shorthand version that makes it feel like a built-in language construct.